Разбивает файл main.js на модули
